# resources/openapi.yml
swagger: "2.0"
info:
  title: BSC API
  description: BSC API
  version: 0.0.1
host: chekov.basecondition.bsc
basePath: /api/v1
...
schemes:
  - https
paths:
...
  /consulting/{id}:
    get:
      tags:
        - consulting
      summary: get existing consulting entry
      description: get existing consulting entry
      operationId: getDate
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          format: "int64"
      produces:
        - application/json
      responses:
        500:
          $ref: '#/responses/500'
        401:
          $ref: '#/responses/401'
        204:
          description: submit success
          schema:
            $ref: '#/definitions/consultingDateGet'
      security:
        - oAuth2: []
    patch:
      tags:
        - consulting
      summary: update existing consulting entry
      description: update existing consulting entry
      operationId: updateDate
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          format: "int64"
      produces:
        - application/json
      responses:
        500:
          $ref: '#/responses/500'
        403:
          $ref: '#/responses/403'
        401:
          $ref: '#/responses/401'
        400:
          $ref: '#/responses/400'
        204:
          description: update success
          schema:
            $ref: '#/definitions/consultingDateAdd'
      security:
        - oAuth2: []
    delete:
      tags:
        - consulting
      summary: delete existing consulting entry
      description: delete existing consulting entry
      operationId: deleteDate
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          format: "int64"
      produces:
        - application/json
      responses:
        500:
          $ref: '#/responses/500'
        403:
          $ref: '#/responses/403'
        400:
          $ref: '#/responses/400'
        204:
          description: update success
      security:
        - oAuth2: []
...
  /structure/events:
    get:
      tags:
        - structure
      summary: get main event categories
      description: get main event categories
      operationId: getEventMainCategories
      produces:
        - application/json
      responses:
        500:
          $ref: '#/responses/500'
        401:
          $ref: '#/responses/401'
        200:
          description: returns list of events
          schema:
            type: array
            items:
              $ref: '#/definitions/eventCategories'
...
securityDefinitions:
  basicAuth:
    type: basic
  apiKey:
    type: apiKey
    in: header
    name: X-API-Key
  oAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://chekov.basecondition.bsc/api/v1/token
    tokenUrl: https://chekov.basecondition.bsc/api/v1/token
    scopes:
      read: Grants read access
      write: Grants write access
      admin: Grants read and write access to administrative information
responses:
  500:
    description: internal_server_error
    schema:
      $ref: '#/definitions/error'
...
  403:
    description: forbidden
    schema:
      $ref: '#/definitions/error'
  401:
    description: invalid_token
    schema:
      $ref: '#/definitions/error'
  400:
    description: invalid_request
    schema:
      $ref: '#/definitions/error'
definitions:
  error:
    type: object
    properties:
      error:
        type: string
      error_description:
        type: string
...
  consultingDateGet:
    type: object
    properties:
      id:
        type: integer
        description: id of consulting menu item
      title:
        type: string
        description: title of consulting menu item (= menuname)
      date:
        type: string
        description: format DD.MM.YYYY
      time:
        type: string
        description: format HH\:II
      location:
        type: string
        description: location selection
      consultant:
        type: string
        description: consultant name
      notice:
        type: string
        description: free notice
      article:
        type: object
        description: article structure
        properties:
          type:
            type: string
            description: should be "info-card" always
          id:
            type: integer
            description: id of article
          value:
            type: object
            properties:
              menuname:
                type: object
                properties:
                  value:
                    type: string
                    description: name of the bubble
              title:
                type: object
                properties:
                  value:
                    type: string
                    description: headline of info-card
              desc:
                type: object
                properties:
                  value:
                    type: string
                    description: description of info-card
              facts:
                type: array
                description: pairs of label, value
                items:
                  properties:
                    label:
                      type: string
                    value:
                      type: string
              overlay:
                type: object
                properties:
                  value:
                    type: string
                    description: overlay content (when clicking request button)
  consultingDateAdd:
    type: object
    properties:
      id:
        type: integer
        description: id of consulting menu item
      title:
        type: string
        description: title of consulting menu item (= menuname)
      date:
        type: string
        description: format DD.MM.YYYY
      time:
        type: string
        description: format HH\:II
      location:
        type: string
        description: location selection
      consultant:
        type: string
        description: consultant name
      notice:
        type: string
        description: free notice
  eventCategories:
    type: object
    items:
      properties:
        name:
          type: string
          description: name of main category
...